requirements.txt
python-dotenv
transformers
tensorflow
tf-keras
pillow
langchain
langchain_community
IPython
openai==0.28
streamlit
